+++ src/Sulu/Bundle/CategoryBundle/Controller/CategoryController.php
@@ -12,7 +12,6 @@
 namespace Sulu\Bundle\CategoryBundle\Controller;
 
 use FOS\RestBundle\Controller\Annotations\Get;
-use FOS\RestBundle\Controller\Annotations\Put;
 use FOS\RestBundle\Routing\ClassResourceInterface;
 use Hateoas\Representation\CollectionRepresentation;
 use Sulu\Bundle\CategoryBundle\Category\CategoryListRepresentation;
@@ -68,16 +67,18 @@ class CategoryController extends RestController implements ClassResourceInterfac
      *
      * @Get("categories/fields")
      *
+     * @param Request $request
+     *
      * @return mixed
      */
-    public function getFieldsAction()
+    public function getFieldsAction(Request $request)
     {
         // default contacts list
         return $this->handleView(
             $this->view(
                 array_values(
                     array_diff_key(
-                        $this->getManager()->getFieldDescriptors(),
+                        $this->getManager()->getFieldDescriptors($this->getLocale($request)),
                         [
                             'depth' => false,
                             'parent' => false,
@@ -295,16 +296,15 @@ class CategoryController extends RestController implements ClassResourceInterfac
 
         $listBuilder = $factory->create(self::$entityName);
 
-        $restHelper->initializeListBuilder(
-            $listBuilder,
-            $this->getManager()->getFieldDescriptors()
-        );
+        $fieldDescriptors = $this->getManager()->getFieldDescriptors($this->getLocale($request));
 
-        $listBuilder->addSelectField($this->getManager()->getFieldDescriptor('depth'));
-        $listBuilder->addSelectField($this->getManager()->getFieldDescriptor('parent'));
-        $listBuilder->addSelectField($this->getManager()->getFieldDescriptor('hasChildren'));
+        $restHelper->initializeListBuilder($listBuilder, $fieldDescriptors);
 
-        $listBuilder->addGroupBy($this->getManager()->getFieldDescriptor('id'));
+        $listBuilder->addSelectField($fieldDescriptors['depth']);
+        $listBuilder->addSelectField($fieldDescriptors['parent']);
+        $listBuilder->addSelectField($fieldDescriptors['hasChildren']);
+
+        $listBuilder->addGroupBy($fieldDescriptors['id']);
 
         if ($parentKey !== null) {
             $this->addParentSelector($parentKey, $listBuilder);